Refactored and added classes for new Data structures
