+++ src/Data/Game.py
@@ -1,0 +1,2 @@
+# Game class - represents the current game. Will hold the two teams, event methods to modify player / team information,
+# and get information from the player / team classes to relay to the displays
