+++ src/Data/Player.py
@@ -1,0 +1,1 @@
+# Player class - represents players in the game. Will contain player score, name (first and last), and codename
