+++ src/Data/Team.py
@@ -1,0 +1,2 @@
+# Team class - represents a team in the game. Will contain a list of players on the team, team score,
+# and other necessary information
